+++ app/src/main/java/woowacourse/payments/domain/model/CardCompany.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.common.model
+package woowacourse.payments.domain.model
 
 import androidx.annotation.DrawableRes
 import androidx.compose.ui.graphics.Color
